services/vertexai/requirements.txt
googleapis-common-protos
grpcio
grpcio-tools
grpcio-reflection
google-cloud-aiplatform
